requirements.txt
streamlit>=1.25
openai>=0.28.0
requests>=2.28
python-dotenv>=1.0
